Widening the fields was right and widening the parameter was not
`instanceof` published at **0.05x** -- the Java reference at 956 us, level with
node's 978, while this lane sat unmoved at 50. A reference twenty times slower
than the thing it measures is not a stricter reference, it is a broken one.

Two declarations changed when the rule only covered one. `javap` says this
backend emits `double size`, `double radius`, `double side`, so the *fields*
were right to widen -- that was the point. The *parameter* went with them, and
`i % 3` became a `drem`, which HotSpot turns into nothing cheap: a probe for its
cost timed out during warmup.

    with `shape(double i)`   956.51 us    0.05x
    with `shape(int i)`       46.52 us    1.08x

1.08x is what the row read this morning, so the corrected reference is the
honest one and nothing about this lane changed.

The failure is worth more than the fix. I was auditing for references that
flattered this compiler, found one that *penalised* it -- `int` fields where we
emit doubles -- corrected that, and overshot into penalising the reference
instead. **"Whatever the subject actually carries" is a question to ask of each
declaration, not a policy to apply to a file.** The fields and the loop counter
are two different questions and I answered them with one edit.

It also says something about the rule's own asymmetry: an unfair reference that
costs us a row is invisible to every instinct tuned for flattery, and the
correction for it lands in the direction those instincts do not check either.

+++ benches/cases/instanceof/ref.java
@@ -27,6 +27,17 @@
 // reason this one does not. The rule is not "always f64"; it is "whatever the
 // subject actually carries", and the only way to know is to look.
 //
+// **The rule is about the fields and stops there.** A first pass widened the
+// *parameter* too -- `shape(double i)` -- and `i % 3` became a `drem`, which
+// HotSpot turns into nothing cheap. The row went from 1.08x to **0.05x**: the
+// reference at 956 us, level with node's 978, while this lane sat unmoved at
+// 50. A reference twenty times slower than the thing it measures is not a
+// stricter reference, it is a broken one -- and it fails the very rule that
+// motivated the change, a cost in one lane only, this time the reference's.
+//
+// So: `double` where `javap` says the class holds a double, `int` for a loop
+// counter, and the two are separate questions.
+//
 // Each subclass adds its own field on top of the base's, which is what
 // `Layout.base` records and what makes `super_class` real here rather than a
 // prefix coincidence.
@@ -57,7 +68,7 @@ final class Ref {
         }
     }
 
-    private static Shape shape(double i) {
+    private static Shape shape(int i) {
         if (i % 3 == 0) {
             return new Circle(i);
         }
